Fix sequencer cache eviction policy
The sequencer cache implements a sliding window to help with cache
eviction, which incorrectly assumes that conflict keys are unique.
This can lead to a memory leak when the same keys are written over
and over again. This patch also fixes a bug where the cache evicts
partial conflicts keys for a particular timestamp. This can lead to
incorrect resolutions and can cause data loss.